Simplify saving model results.
Major changes to results saving code. Includes saving the initialization data
as timestep zero (with events data, with does not apply to T0 marked as nan in
the output). Also the LULC results is no longer saved separately from the
population and events data, simplifying handling it in R or other stats
software.